fixed bugs with user login and registration
