+++ eCommerce/Service/Security.cs
@@ -5,9 +5,8 @@ using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
-using  eCommerce;
 
-namespace eCommerce
+namespace eCommerce.Service
 {
     internal class Security
     {
